feat: support returning clause in delete statement
