+++ crates/uroborosql-fmt/test_normal_cases/dst/064_delete_returning.sql
@@ -1,0 +1,8 @@
+delete
+from
+	products
+where
+	obsoletion_date	=	'today'
+returning
+	*
+;
