+++ crates/uroborosql-fmt/test_normal_cases/src/064_delete_returning.sql
@@ -1,0 +1,3 @@
+DELETE FROM products
+  WHERE obsoletion_date = 'today'
+  RETURNING *;
